requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
feedparser==6.0.10
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.3
